separate errors in a single file
